+++ CHANGELOG.md
@@ -8,6 +8,10 @@ Please add one entry in this file for each change in Yarn's behavior. Use the sa
 
 ## 1.12.0
 
+- Adds `yarn global outdated`
+
+  [#6488](https://github.com/yarnpkg/yarn/pull/6485) - [**Matthias Winkelmann**](https://github.com/MatthiasWinkelmann)
+
 - Adds initial support for PnP on Windows
 
   [#6447](https://github.com/yarnpkg/yarn/pull/6447) - [**John-David Dalton**](https://twitter.com/jdalton)